requirements.txt
# Binance API
python-binance==1.0.19
ccxt>=4.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2

# Async & WebSocket
websockets==12.0
aiohttp==3.9.1

# Scheduling
schedule==1.2.0
APScheduler==3.10.4

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
